requirements.txt
imutils==0.5.4
numpy==1.24.4
opencv-python==4.9.0.80
pillow==10.3.0
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pyparsing==3.0.6